update: VaccinationRecord -> VaccineRecord로 파일명 변경 및 id개수로 체크 박스 생성
